+++ packages/orq-rc/src/orq_ai_sdk/models/filelistop.py
@@ -74,7 +74,7 @@ class FileListData(BaseModel):
     workspace_id: str
     r"""The id of the resource"""
 
-    created: Optional[datetime] = parse_datetime("2025-09-02T12:49:02.206Z")
+    created: Optional[datetime] = parse_datetime("2025-09-03T05:07:29.348Z")
     r"""The date and time the resource was created"""