+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasourceop.py
@@ -100,7 +100,7 @@ class UpdateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01K457JMT7KKEKNNC1H90B859N"
+        "01K46ZJ83YZ35BPS7G9TB4W2AP"
     )
     r"""The id of the resource"""