+++ packages/orq-rc/RELEASES.md
@@ -2509,3 +2509,13 @@ Based on:
 - [python v3.12.0-rc.27] packages/orq-rc
 ### Releases
 - [PyPI v3.12.0-rc.27] https://pypi.org/project/orq-ai-sdk/3.12.0-rc.27 - packages/orq-rc
+
+## 2025-09-03 05:09:04
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.607.0 (2.690.1) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v3.12.0-rc.28] packages/orq-rc
+### Releases
+- [PyPI v3.12.0-rc.28] https://pypi.org/project/orq-ai-sdk/3.12.0-rc.28 - packages/orq-rc